// src/lib/types/transportTypes.ts
export enum TransportStatus {
  PENDING = "PENDING",
  TRUCK_ASSIGNED = "TRUCK_ASSIGNED",
  EN_ROUTE_TO_WAREHOUSE = "EN_ROUTE_TO_WAREHOUSE",
  ARRIVED_AT_WAREHOUSE = "ARRIVED_AT_WAREHOUSE",
  LOADING = "LOADING",
  LOADING_COMPLETED = "LOADING_COMPLETED",
  EN_ROUTE_TO_DESTINATION = "EN_ROUTE_TO_DESTINATION",
  ARRIVED_AT_DESTINATION = "ARRIVED_AT_DESTINATION",
  UNLOADING = "UNLOADING",
  UNLOADING_COMPLETED = "UNLOADING_COMPLETED",
  DELIVERED = "DELIVERED",
}